+++ src/Commands/HorusCommandType.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace HorusMod.Commands
+{
+    public enum HorusCommandType
+    {
+        SpawnUnit,
+        SpawnGroup,
+        DeleteUnit,
+        FactoryAction,
+        EconomyAction
+    }
+}